Plagiarism detection System/requirements.txt
flask
scikit-learn
pypdf
python-docx
